Clean up PhoneConfirmation concern
The method was only used in PhoneSetupController, so move method & private method there and stop
including this where it is not used.
