+++ app/controllers/sign_up/registrations_controller.rb
@@ -1,6 +1,5 @@
 module SignUp
   class RegistrationsController < ApplicationController
-    include PhoneConfirmation
     include ApplicationHelper # for ial2_requested?
 
     before_action :confirm_two_factor_authenticated, only: [:destroy_confirm]
