+++ app/controllers/users/piv_cac_setup_controller.rb
@@ -1,6 +1,5 @@
 module Users
   class PivCacSetupController < ApplicationController
-    include PhoneConfirmation
     include ReauthenticationRequiredConcern
 
     before_action :confirm_two_factor_authenticated
